% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Count non-infinit values}
\usage{
counter(M)
}
\arguments{
\item{M}{is a matrix.}
}
\value{
Returns total number of non-infinit values.
}
\description{
\code{counter} is used to get total number of non-infinit values in a matrix.
}
\details{
The input of this function is a matrix. This function returns the total number of non-infinite values
in the given matrix. In order to work properly, \code{M} must be a \code{\link{matrix}} otherwise this method
generates an error.
}
\examples{
a <- matrix(data = c(2, Inf, Inf, 0, Inf, Inf, Inf, 10, Inf), nrow = 3, ncol = 3, byrow = TRUE)

counter(a)

# [1] 3

}
